MINIFICPP-1445 - Refactor docker integration test frame
