+++ docker/test/integration/minifi/core/Processor.py
@@ -1,0 +1,45 @@
+from .Connectable import Connectable
+
+class Processor(Connectable):
+    def __init__(self,
+                 clazz,
+                 properties=None,
+                 schedule=None,
+                 name=None,
+                 controller_services=None,
+                 auto_terminate=None):
+
+        super(Processor, self).__init__(name=name,
+                                        auto_terminate=auto_terminate)
+
+        if controller_services is None:
+            controller_services = []
+
+        if schedule is None:
+            schedule = {}
+
+        if properties is None:
+            properties = {}
+
+        if name is None:
+            pass
+
+        self.clazz = clazz
+        self.properties = properties
+        self.controller_services = controller_services
+
+        self.schedule = {
+            'scheduling strategy': 'TIMER_DRIVEN',
+            'scheduling period': '1 sec',
+            'penalization period': '30 sec',
+            'yield period': '1 sec',
+            'run duration nanos': 0
+        }
+        self.schedule.update(schedule)
+
+    def nifi_property_key(self, key):
+        """
+        Returns the Apache NiFi-equivalent property key for the given key. This is often, but not always, the same as
+        the internal key.
+        """
+        return key
