+++ docker/test/integration/minifi/core/RemoteProcessGroup.py
@@ -1,0 +1,13 @@
+import uuid
+
+class RemoteProcessGroup(object):
+    def __init__(self, url,
+                 name=None):
+        self.uuid = uuid.uuid4()
+
+        if name is None:
+            self.name = str(self.uuid)
+        else:
+            self.name = name
+
+        self.url = url
