+++ docker/test/integration/minifi/processors/ListenHTTP.py
@@ -1,0 +1,14 @@
+from ..core.Processor import Processor
+
+class ListenHTTP(Processor):
+    def __init__(self, port, cert=None, schedule=None):
+        properties = {'Listening Port': port}
+
+        if cert is not None:
+            properties['SSL Certificate'] = cert
+            properties['SSL Verify Peer'] = 'no'
+
+        super(ListenHTTP, self).__init__('ListenHTTP',
+			properties=properties,
+			auto_terminate=['success'],
+			schedule=schedule)
