+++ docker/test/integration/minifi/validators/SegfaultValidator.py
@@ -1,0 +1,8 @@
+from .OutputValidator import OutputValidator
+
+class SegfaultValidator(OutputValidator):
+    """
+    Validate that a file was received.
+    """
+    def validate(self):
+        return True
