+++ docker/test/integration/test_hash_content.py
@@ -14,8 +14,6 @@
 # limitations under the License.
 
 from minifi import *
-from minifi.test import *
-
 
 def test_hash_invoke():
     """
