+++ docker/test/integration/test_rdkafka.py
@@ -14,8 +14,6 @@
 # limitations under the License.
 
 from minifi import *
-from minifi.test import *
-
 
 def test_publish_kafka():
     """
