+++ docker/test/integration/test_s3.py
@@ -14,8 +14,6 @@
 # limitations under the License.
 
 from minifi import *
-from minifi.test import *
-
 
 def test_put_s3_object():
     """
